+++ NimbusBackend/src/test/java/nl/nimbus/backend/NimbusBackendApplicationTests.java
@@ -1,4 +1,4 @@
-package nl.nimbus.nimbusBackend;
+package nl.nimbus.backend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
