--- nimbusBackend/src/main/java/nl/nimbus/nimbusBackend/NimbusBackendApplication.java
@@ -1,13 +1,0 @@
-package nl.nimbus.nimbusBackend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NimbusBackendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(NimbusBackendApplication.class, args);
-	}
-
-}
